Split secrets:edit into secrets:setup and secrets:edit.
Also add a general secrets command that does setup first, then
edits.

Required some more finagling to the command infrastructure.
